Move connection logger to internal/utils
Suggested-by: Leo Chen <leo@harmony.one>
